Move the eager vm save to enterTCHelper
Summary: We always save rVmFp and rVmSp as part of the service
request stub; but we could just as easily do it in
enterTCHelper, and reduce the size of the stubs.

When I went to reduce the size of the stubs, I saw that we
were already somewhat undercounting - although we never come
close to the worst case in practice. So maxStubSpace is
actually unchanged by this diff (although the actual space
for non-ephemeral stubs will be reduced).

Reviewed By: @jdelong

Differential Revision: D1632905