Enable -fglobal-dce and -fremove-unused-locals in hhbbc by default
Summary: For some reason I forgot that I meant to look at doing this when
turning off postOptimize.  The pass was off originally because I wanted to
think about it some more and was worried about the default implementation being
correct for bytecodes that aren't explicitly supported---it really just relies
on the mayReadLocalSet being correct from the normal abstract interpreter,
which I feel pretty confident about now.

This also turns on unused local variable removal, so we don't need to do
-fno-insert-assertions on test/quick/lotsa-locals.php (because hhbbc just
removes all the locals after constant propagating them into var_dump).  Found
and fixed a bug in this---it was willing to remove parameter locals.

Reviewed By: @jano

Differential Revision: D1633779