Remove isRefCounted; fix a bug in simplifier xorBoolImpl
Summary: isRefCounted really means maybeCounted, which we have in Type.  And in
xorImpl we needed to check that the sources can't be doubles, not that they
aren't known to be doubles.  (As alexanderm pointed out in the task.)

Disable it for arrays and objects even though we don't have an example it
breaks, because the burden of proof should be on it being correct.

Reviewed By: @alexmalyshev

Differential Revision: D1633947