Initial foundation for merging CodeGenerator classes
Summary: This is a real idea, but if people hate it I can pursue something else.

The plan is to gradually convert as many opcodes as possible in the ARM
class to DELEGATE_OPCODE first, before eventually merging the classes.
This will make the eventual merge go as smoothly as possible. Here's how
I envision it:

- Turn as many opcodes as possible into DELEGATE_OPCODE on the ARM side,
  both ones that are currently punting and ones that have real
  implementations. This will probably entail adding an x64->arm lowering
  pass to allow to use more of the x64 implementations unmodified.

- When that's done, merge the classes. Everything will go through
  x64::CodeGenerator. For the opcodes that still have divergent ARM
  implementations, we'll create things like cgBlahX64 and cgBlahARM,
  mechanically

- For those divergent opcodes, we can gradually replace them with mostly
  merged versions, with a little bit of branching in the middle.

Reviewed By: @edwinsmith, @swtaarrs

Differential Revision: D1635576