Always use RuntimeOption::SourceRoot to resolve repo-auth lookups
Summary: SourceRootInfo::GetCurrentSourceRoot() is used to find the
DocumentRoot, and in RepoAuthoritative mode its also used to specify
the root relative to which all the files were compiled.

Normally, in RepoAuthoritative mode, the DocumentRoot would be the same
as RuntimeOption::SourceRoot, and so everything seems to work.

In FastCGI mode, the transport can override the DocumentRoot; but we
still need to lookup files in the repo relative to the SourceRoot.

Reviewed By: @edwinsmith

Differential Revision: D1638371