Add remaining CALL_OPCODEs for ARM
Summary: In theory, anything that's a CALL_OPCODE on x64 should be the same on
ARM and it should work without problems. I just copied the whole
CALL_OPCODE section from code-gen-x64.cpp and it seems to have worked.
(That resulted in a lot of reordering, so the best way to see what's
changed is to concentrate on the removed PUNT_OPCODEs.)

Reviewed By: @edwinsmith

Differential Revision: D1637822