Update the typechecker's auto-import list for the "HH" namespace
Summary: Ari encountered a bug when trying to declare an async function with a
return typehint inside a namespace. The typechecker did not recognize that
"Awaitable" was auto-imported from the HH namespace, and this was causing
problems.

This diff brings the typechecker's auto-import list up to date, excluding
"IMemoizeParam" which has issues that will be investigated later.

Reviewed By: @jwatzman

Differential Revision: D1643034

Signature: t1:1643034:1414539326:9597edc165f9b1a11087ab5ddc7f53390465bbc7