Make autocomplete work with nullsafe operator
Summary: The issue is that the nullsafe operator doesn't work for properties.

$thing->AUTO332 (the autocomplete token) looks like a property.

We were stopping in naming if we had ?-> accessing a property. This makes it add the error, but now it also keeps going now, so that autocomplete works.

As a side effect, we'll get more errors now if accessing a property also produces a typing error.

In a future diff, I'll make the nullsafe operator autocomplete not include properties.

Reviewed By: @jwatzman

Differential Revision: D1630608