Fix some problems in Type::operator-
Summary: Type::operator- was potentially removing valid values from the result.
E.g. Int - Int<24> or Obj - Obj<Foo> would return Bottom, or
Array=(MapLikeConstant) - Array(RATPacked) would return Bottom.  This could
easily do something incorrect (if that type flowed to the wrong place, we could
e.g. fold a CheckType<Int> branch or something incorrectly based on it).  It
also was sometimes going the other way: if you subtracted a specialized array
from another specialized array, you'd get back Type::Arr, which is
unnecessarily too big (we know *this - x can't ever be any bigger than *this).

I was a little bit afraid about how we might be using this stuff---but an audit
found no uses where the right hand side wasn't a constant type that was
unspecialized (it's mostly about removing Null or Uninit from things, so I
think we're likely to be ok).  I would rather have function called
"remove_null" (or whatnot) do that sort of thing so it's clear what we're
trying to do, though---this function seems too easy to misuse...  It is
incorrect to use it with arbitrary types that you didn't control if you aren't
ok with a - b returning a after this diff (and before this diff it's incorrect
to use it in that situation because it aborted on some inputs).

Reviewed By: @markw65

Differential Revision: D1645584