Convert built-in PDO class to HNI
Summary: **Motivation**
The Doctrine libary used by Symfony defines a class that extends the built-in PDO and implements their custom interface. That interface declares the `prepare(..)` method with one parameter, while `PDO`'s version has two, the second having a default value.
With the class defined via IDL, the semantic check in `class.cpp` was failing to see the presense of the default value and reporting an error.

One of Symfony's test that were failing:
  Symfony/Bridge/Doctrine/Tests/Form/ChoiceList/UnloadedEntityChoiceListCompositeIdTest.php

Reviewed By: @ptarjan

Differential Revision: D1612690