Relax parameter declaration check for built-in functions
Summary: In certain cases we can't have our built-ins act as those of PHP5 because the latter don't have type hints.
For example, consider the following case:

  // Built-in PDO (HHVM)
  class PDO {
   . . .
   function query(string $sql, ...$argv): mixed;
   . . .
  }

  // Built-in PDO(PHP5)
  class PDO {
   . . .
   function query();
   . . .
  }

  interface I {
   function query();
  }

  class C extends PDO implements I {
  }

This code works in PHP5, but doesn't in HHVM, because our parameters check fail.

The proposed solution is to disable parameter checks for built-in classes when not running in HipHop-syntax mode.

Reviewed By: @markw65

Differential Revision: D1623635