Make emitProfiledGuard not use constrainValue on PtrToFoo types
Summary: We need to stop using constrainValue on pointer types if we want inner
types to be sound (because almost all of them will be PtrToGen).  This is an
easy one---it already knows which local or stack slot it's dealing with, and
moreover if it uses the ldLocAddr and ldStackAddr helpers it'll do the
constrain for it.  Because of preOptimizeLdLocAddr it's still going to use bad
pointer types for the locals for now, though.

This diff also makes one small behavior changes (believed to be a bug fix):
emitProfiledGuard in checkTypeStack was adding evalStack.size() instead of
evalStack.numCells(), which I think means it could guard the wrong offset, but
I think in practice we always do this when the stack has been spilled so it
doesn't happen.

Reviewed By: @alexmalyshev

Differential Revision: D1650794