Turn off MemeLim store sinking for now (it isn't right)
Summary: This is currently willing to move stores to places where the value
that would've been in the memory location is different.  We need to limit store
motion to blocks dominated by the old store, not just dominated by the value
that was stored, or do a more general version of this.

Reviewed By: @markw65

Differential Revision: D1654195