Fribidi extension for HHVM
Summary: Kind of a noob to this stuff but this is a straight port of pecl-fribidi.  I have some concerns over how the strings are being managed...  for a PHP extension you can specify whether or not the string being returned should be copied.  Not sure how to do that for HHVM.

Reviewed By: @ptarjan, @behnam

Differential Revision: D1632647