More fixes to kernel version pattern
Summary: Restrict the pattern to match the appropriate number of characters,
and assert that an acceptable number of fields are parsed.  Also split up the
test cases so it's easier to see which ones are failing.

Reviewed By: @markw65

Differential Revision: D1655782

Signature: t1:1655782:1415049423:1e1a6f665c1c8f129e48f9df120f7715f69034ca