More pretty printing fixup
Summary: Vasm code gen can generate blocks in the cold or frozen regions
even if the particular IR block is not hinted at being in either. The
pretty printed disassembly step at the end of code gen did not account
for this, so it would only print the code from the area that the IR
block originally though it was in. This diff fixes this issue by iterating all
of the areas for each IR instruction and dumping all of the code from any asm
range associated with that instruction.

Reviewed By: @jdelong

Differential Revision: D1655171