Thread-safe LRU PCRE cache
When the PCRE cache fills up, memory is leaked since the static string
created by insert_cached_pcre() cannot be freed. Also, queueing cached
regexes for deletion on request termination is not good policy -- it
is effectively a memory leak for long-running scripts.

Also, when the PCRE cache is full, new regexes need to be recompiled on
every call to preg_match() etc. so the performance of an app using
dynamically-generated regex strings will degrade once the limit is
reached.

* Developed a new thread-safe LRU container class, published to
  https://github.com/tstarling/thread-safe-lru and copied here. Apache
  license chosen to allow future inclusion into Folly if necessary.
* Made container class choice a runtime option, with two new classes
  plus the old AtomicHashArray solution.
* Reorganised PCRE cache code, made it OOP.
* In the AtomicHashArray case, use reference counting to destroy
  temporary objects, instead of deferring destruction until the end of
  the request. Remove m_overflow-related housekeeping.