Convert easy arithmetic ops to DELEGATE
Summary: My plan for now is to convert everything I can that's explicitly
implemented in code-gen-arm.cpp before tackling the long list of punting
opcodes.

Another few issues here, most notably `imul`. I think emitting `imul` as
`Mul` on ARM is wrong, because I realized we're actually counting on it
to set condition flags, and there's no multiply instruction on ARM that
will do that. I split it into another vasm opcode to make this clear.

It'd be nice to use non-flags-setting ARM opcodes when they're not
needed; I guess this can be done in a later pass (after DCE determines
that the flags aren't used).

Reviewed By: @edwinsmith

Differential Revision: D1658861