Basic parser for AT
Summary: This is a simple parser for Box's augmented types, as the
basis of an AT-to-Hack conversion tool. It was optimized for being easy
to write, not really for having good error recovery/detection or for
being super flexible, since AT isn't really changing, and this will be a
run-once tool.

https://github.com/box/augmented_types/wiki/The-Type-System

Reviewed By: @ptarjan

Differential Revision: D1652771

Signature: t1:1652771:1415154622:03f5f990895cedb819126df07dc6c439d898f510