Add support for pre-built branches to benchy
Summary: Many times you will have already built a branch somewhere and you
want to use that build rather than having to rebuild everything to a place
that benchy knows about. This patch adds the ability to specify the location
of a pre-built VM so that this workflow is now possible.

This diff also updates the tab complete script to support tab completion
of paths as part of the VM. Run with -h to see an example of how to use
it.

Reviewed By: @jdelong

Differential Revision: D1646858