Make integration test restore state at a different root directory
Summary: Also modify the way we check test results -- instead of stripping off the root prefix of the file paths in the test output, we fill in what the root should be in our expected output. Thus the test ensures that we are not saving the wrong absolute paths to disk.

Reviewed By: @jwatzman

Differential Revision: D1654589

Signature: t1:1654589:1415139391:64e0e1afc254878a7a0fdc6b19f733897f57c41c