version of memcpy-x64 optimized for the tail end of 7,6,5,4 bytes in a N byte copy
Summary: I see a performance win of 2-3% on wordpress-3.6 with this version of memcpy. In this particular case I have optimzed the memcpy for the tail end of 7,6,5,4 bytes when a N byte copy is requested.

I have another revision where I optimized all the cases for <7 bytes and <63 bytes tail ends however I see an increase on L1 cache misses, with it.
Please review this on the FB workload. If a histogram is available for the byte copies that take place I could specifially optimize it for those cases.

Closes https://github.com/facebook/hhvm/pull/4137

Reviewed By: @markw65

Differential Revision: D1661080

Signature: t1:1661080:1415295604:debcb26bfd4075434400368b77463edd411b3727

Pulled By: @sgolemon