Fix branch shrinking code
Summary: The branch shrinking code assumed that if it was
possible to shrink a branch in the source of the relocation,
then it was possible in the destination. Thats not true because
some instructions need special treatment at page boundaries,
which can cause nops to be inserted.

The solution is to iterate, marking the failed ones wide. Since
its possible that a branch was already emitted in narrow form by
code gen, and it may no longer be possible after relocation,
also add code to widen branches when necessary.

This exposed bugs in the rewind code, when an exception is thrown
from relocate, so fix those too.

Reviewed By: @edwinsmith

Differential Revision: D1664748