Add an intrinsic to disable inlining for specific functions
Summary: Disabling inlining of specific functions can be useful when
trying to write reliable, reproducible test cases. This diff adds the
ability to disable inlining for particular functions through the use of
the disable_inlining intrinsic, part of a new __hhvm_intrinsics
StandardExtension. The extension is hidden so that it won't be loaded
unless enabled with the Eval.EnableIntrinsicsExtension runtime option.

Reviewed By: @jdelong

Differential Revision: D1659302