Move MInstrState from C++ stack to RDS header
Summary: Move MInstrState to be with the other VM registers in the RDS header,
rather than reserving space for it on the native stack in enterTCHelper. This
is one step on the way to allowing arbitrary rsp manipulations during
tracelets. I also removed MInstrState's tvVal and tvResult members since
they've been unused for a while.

Reviewed By: @jasone

Differential Revision: D1665434