Provide mechanism to "blacklist" tests that must be run serially
Summary: @​swtaarrs noticed that `ext_memcached` tests couldn't be run in parallel because
of key collisions. So, we now have a way to make sure they (and other tests that
we deem "parallel challenged") to always run serially.

Note: This could slow the test runner down just a bit, of course, when running these
tests. However, with memcached off, there is not a really big noticeable difference. Right now I found 5 memcached tests that have to be run this way. The slowness will be more
noticeable with the fewer tests you run that include the serial ones, since we have to
reduce the number of threads. However, for large test sets, there should be minimal
differences since we maxed the number of threads anyway.

Fixes #3884

Reviewed By: @paulbiss

Differential Revision: D1666135

Signature: t1:1666135:1415344877:5bfdad4d0bf63ed3019f17846d66b34ea35782e1