Allow mysqli objects to be passed into the mysqli_result constructor
Summary: Currently we only allow resources to be passed to `mysqli_result::__construct`. Apparently there is an undocumented feature that allows `mysqli` objects to also be passed in as well. In fact, it seems PHP 5.x pretty much allows anything to be passed to the constructor and it will just warn that properties aren't available and move on. I didn't go that far.

Fixes #2082

Reviewed By: @paulbiss

Differential Revision: D1663326

Signature: t1:1663326:1415332481:de9b5479c64e476b8801efe4ae3bcb84698fcfcf