Side exit on surprise flags for backward jumps
Summary: We handle surprise flags on function entry and exit, and need to for
the profiler to be reasonable, but it doesn't look like this one is needed for
that.  But having the surprise flag check rejoin the main line on back edges
seems like it will unnecessarily pessimize some possible code transformations
we may want in the future.  Right now, SupriseHook calls a function which can
invoke arbitrary php (without throwing), because it handles any pending
signals.  This means any potentially-escaped heap state we may want to track is
going to need to be thrown away at every back edge, which seems like we
shouldn't do.

Reviewed By: @ottoni

Differential Revision: D1669109

Signature: t1:1669109:1415503777:93d3b94947e60de2786ff6114a2a010827dee175