Make profiled packed array gets not rejoin the main line
Summary: When we emit a packed array get, we're generating a control flow
diamond that does a generic ArrayGet in astubs.  Since we only do this if it
was purely monomorphic during profiling, this seems like a bummer.  It also
means we can't propagate a refineLocalValue on the CheckType past the merge
point, and duplicate checks on the same array being packed can occur.

Reviewed By: @swtaarrs

Differential Revision: D1669141

Signature: t1:1669141:1415648447:e199d21362cc36b30429e4580db571208b42b36f