ftell() for sockets shouldn't be supported. So always return 0.
Summary: Right now, for sockets, we return the number of bytes written to the socket when we call
`ftell()`. But this shouldn't be the case.

See ESPIPE ... http://pubs.opengroup.org/onlinepubs/009695399/functions/ftell.html

PHP 5.x returns 0 when calling `ftell()` on a socket, so that's what I propose we doo.

This is a one line change, which can be the scariest. :)

Let me know if I am off base, but this does fix an HHVM OSS issue.

Fixes #3018

Reviewed By: @sgolemon

Differential Revision: D1668651

Signature: t1:1668651:1415413974:5c040e68e2b4ca1f7a206eef6eca35f54eac12b0