Improve memory_effects, make it talk in terms of abstract memory locations
Summary: Also teaches it to distinguish between may/must information for both
loads and stores, so it can be used for more than just DSE.  Audited all
instructions for effects on array elements and object properties, but a few
cases are still probably overly conservative for now (it's not clear to me that
we'll need to give up on array elements ever for instructions that might
re-enter).  This diff updates memelim to use the new interface.

Reviewed By: @edwinsmith

Differential Revision: D1669287