Drop FrameState's ref prediction if type grows to Gen.
Summary: Was hitting this with fixed-point. A value was Arr in one path, and
BoxedArr in the other, and we ended up with a state of
type=Gen,predict=BoxedArr. This later broke upon hitting a CheckLoc<Arr>
as we refined Gen -> Arr, but kept around the BoxedArr prediction.

Add the invariant to the field's comment and checkInvariants.

Reviewed By: @ottoni, @jdelong

Differential Revision: D1668687

Signature: t1:1668687:1415415982:3e94da4b41306ba74bbfc6c50a8f77331593f2e7