Wait for clean shutdown on SIGHUP
Summary: HHVM is supposed to do a clean shutdown when it receives SIGHUP -
which it does when using LibEventServer - but with ProxygenServer, it spins off
an additional thread to shutdown queues, leaving the original thread to
re-raise the signal before the queues drain.  Fix this by waiting for the
servers to stop.

Reviewed By: @markw65

Differential Revision: D1675300

Signature: t1:1675300:1415808825:f77150b8cb0ea85eaf90d58095be4baeaea2d317