Add support for syncpoints and exceptions in the llvm backend
Summary: Both are fairly straightforward and mostly involve parsing information
out of the data sections emitted by LLVM. As far as I can tell, LLVM just emits
the exception tables into memory we give it and doesn't actually register any
of the exception information with the runtime. This means we can allocate all
data sections it asks for with malloc and free them once we extract the
information we need, rather than leaking them in globalData.

Reviewed By: @maksfb

Differential Revision: D1678102