Remove bad assert from BaseVector::data()
Summary: BaseVector::data() had an assert that was checking if m_data[0] was a Cell.
This assert is bogus because the Vector could be empty, and in such cases
it is not valid to dereference m_data[0]. It turns out that this assert was
redundant anywhere because the only place that uses BaseVector::data()
already asserts that each element in the Vector is a Cell (also, there are
a bunch of asserts throughout ext_collections that make sure that each
element is a Cell).

Reviewed By: @elgenie

Differential Revision: D1680393

Signature: t1:1680393:1415922724:0331eb4ed100231a24517cf4870736a07f56e275