Add zlib.output_compression ini option
Summary: We had `zlib.output_compression_level`, but that should only be used if
`zlib.output_compression` is set to "On". Add this option.

This doesn't perform the gamut of the option, but it is a start. I have a TODO
to add more functionality for the option.

http://docs.hhvm.com/manual/en/zlib.configuration.php

Fixes #3744

Reviewed By: @paulbiss

Differential Revision: D1677334

Signature: t1:1677334:1415931492:a151c55a03ccb03b83efc54107f598e9b01150e3