Fix problems in getCurrentPropertyOffset relating to m_ni.inputs
Summary: We can crash in here because knownCls can be non-null while baseCls is
null.  This can happen if you have a region containing only "FPushObjMethodD;
FCallD", and the callee is inlinable.  What happens is we'll generate a generic
push, and inline a call where all we know is that the $this is Type::Obj.
(This is probably a "bug" in its own right, but there are multiple things to
fix: one is that we should not generate regions like this (there was an assert
type right before that), another is that FCallD should probably influence the
IR's knowledge of the type of the object it's making a method call on, and a
third is that we should not be using types we determined during region
selection (from NormalizedInstruction::inputs) without knowing why they are
there (were they statically known, or due to a guard, etc).

For now just disable this situation if we don't know the type from this
translation.

Reviewed By: @alexmalyshev

Differential Revision: D1681932