Remove duplicate stack size code
Summary: We already had code forcing the stack size to 8M,
but recently added more code to force it to 8 in normal mode,
and 16 in ASAN mode.

In addition, the old code always ran, even if the main thread wasn't
going to do much real work (eg when running as a server).

Drop the old code in favor of the new.

Reviewed By: @edwinsmith

Differential Revision: D1683119

Signature: t1:1683119:1416055986:e7fecf294752fc6f2f17478f613919f46929eea2