Fix some not-type-safe LdRef code generated in a situation with SetOpL
Summary: SetOpL for concat was passing a nullptr for the ldrefExit, claiming it
was because it already guarded.  But this is reason is suspect, because
ConcatCellCell can call arbitrary code via an object __toString.  With the
nullptr, in trunk we generate a LdRef<Str> without a taken branch on a
BoxedCell that may not contain a Str, which means we have a value in the IR
that is potentially mistyped.  Good news is that since it's immediately
followed by a StRef, we'll never use it (as frame state will immediately wipe
out any tracked copies of it), but it's still not ok.

This diff just fixes the comment and modifies stLocNRC to not load the old
inner cell for this situation, so we don't need the guard for a legitimate
reason.

Reviewed By: @swtaarrs

Differential Revision: D1684096