Separate checking whether a Ctx is a $this pointer from loading it
Summary: Use LdCtx to load a Ctx, and add some instructions to check if it is a
non-null $this pointer or cast it to a particular object type.  This diff
doesn't try to address some refcount opts dependencies on the optimizations
relating to $this---as far as I can tell it should be preserving the existing
behavior over there, but I added some TODOs on simplifier and preOptimize rules
it depends on.

Reviewed By: @ottoni

Differential Revision: D1684191