Add sparse-id-containers support for lookups with non integer types
Summary: First time I tried to use it I wanted to give it an SSATmp* key, which
it didn't support.  Make it support structures that convert to ints by default;
for pointers you need to pass a converter.  In the jit add some type aliases to
make that short and use it at the one use site.

Reviewed By: @edwinsmith

Differential Revision: D1684325

Signature: t1:1684325:1416235658:7685412f24703675d719abaed9701329e623dcf4