Record the start time for every request
Summary: We recently refactored the code to record an access log
even for requests that failed early (ie didn't get processed by
the RequestHandler, or which the RequestHandler decided to abort
early).

The requests that previously didn't get processed at all, also didn't
get their start times set. This fixes that.

Also, the previous change moved the point at which we recorded the start
time (when we did record it) past some early setup - which may have
artificially reduced the time recorded in the logs. This puts it back
before we start doing any work for the request.

Reviewed By: @​marccelani

Differential Revision: D1684356